+++ src/pages/implementation-references/react-hook-form.tsx
@@ -2,12 +2,13 @@ import { NextPage } from 'next';
 import { useEffect } from 'react';
 import { SubmitHandler, useForm, FormProvider } from 'react-hook-form';
 import { TextInput } from '@/components/forms/text-input';
-import { Button } from 'antd';
+import { Button, Form } from 'antd';
 import { FileDropzone } from '@/components/forms/file-dropzone';
 import { FileWithPath } from 'file-selector';
+import { PasswordInput } from '@/components/forms/password-input';
 interface ISampleForm {
-  usuario: string;
-  projeto: string;
+  text: string;
+  password: string;
   file: FileWithPath[];
 }
 
@@ -22,17 +23,22 @@ const ReactHookFormPage: NextPage = () => {
 
   return (
     <FormProvider {...RHForm}>
-      <form
-        onSubmit={RHForm.handleSubmit(onSubmit)}
-        className="w-full flex flex-col space-y-4 justify-center items-center"
+      <Form
+        layout="vertical"
+        onSubmitCapture={RHForm.handleSubmit(onSubmit)}
+        className="w-full flex flex-col space-y-2 justify-center items-center"
       >
-        <TextInput<ISampleForm> name="usuario" placeholder="usuario" label="" />
-        <TextInput<ISampleForm> name="projeto" placeholder="projeto" label="" />
+        <TextInput<ISampleForm> name="text" placeholder="Texto" label="Texto" />
+        <PasswordInput<ISampleForm>
+          name="password"
+          placeholder="Senha"
+          label="Senha"
+        />
         <FileDropzone name="file" />
         <Button type="primary" htmlType="submit">
           Enviar
         </Button>
-      </form>
+      </Form>
     </FormProvider>
   );
 };